% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BUNDESLAND_TO_ISO}
\alias{BUNDESLAND_TO_ISO}
\title{Mapping from Bundesland ISO 3166-2 codes to full names}
\format{
A data frame with 16 rows and 2 variables:
\describe{
  \item{bundesland}{full name of Bundesland}
  \item{bundesland_iso}{ISO 3166-2 code of Bundesland}
  ...
}
}
\source{
\url{https://de.wikipedia.org/wiki/ISO_3166-2:DE}
}
\usage{
BUNDESLAND_TO_ISO
}
\description{
A dataset containing the ISO 3166-2 codes and full names of the German Bundeslaender
}
\keyword{datasets}
