% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_common_path}
\alias{get_common_path}
\title{get common part of path from excel file path}
\usage{
get_common_path(xlsx_path)
}
\arguments{
\item{xlsx_path}{character. path to excel file.}
}
\value{
character. common part for all files related to one download, i.e. {timestamp}_impfmonitoring
}
\description{
get common part of path from excel file path
}
