% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{rki_clean_bundesland}
\alias{rki_clean_bundesland}
\title{clean data}
\usage{
rki_clean_bundesland(.data)
}
\arguments{
\item{.data}{raw data frame of sheet "Presse"}
}
\value{
cleaned data in wide format
}
\description{
cleans column names, removes * from Bundesland column, merges Bundesland ISO code and removes "Gesamt"
and comment rows at the end of the table. Returns one row for each Bundesland.
}
