% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decumulated_data.R
\name{rki_decumulate_data}
\alias{rki_decumulate_data}
\title{decumulate cumulative time series}
\usage{
rki_decumulate_data(cumulative_data)
}
\arguments{
\item{cumulative_data}{tibble. the cumulative time series.}
}
\value{
tibble. "decumulated" time series, i.e. where variables represent the new counts since the previous count.
}
\description{
decumulate cumulative time series
}
