% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_data.R
\name{rki_download_cumulative_ts}
\alias{rki_download_cumulative_ts}
\title{download time series from GitHub}
\usage{
rki_download_cumulative_ts()
}
\value{
tibble. Cumulative time series
}
\description{
Downloads cumulative time series from the `vaccc19de_rki_data` repository, more specifically https://raw.githubusercontent.com/friep/vaccc19de_rki_data/main/data/cumulative_time_series.csv
}
