% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decumulated_data.R
\name{rki_download_decumulated_ts}
\alias{rki_download_decumulated_ts}
\title{download decumulated time series from GitHub}
\usage{
rki_download_decumulated_ts()
}
\value{
tibble. 'Decumulated' time series
}
\description{
Downloads 'decumulated' time series from the `vaccc19de_rki_data` repository, more specifically https://raw.githubusercontent.com/friep/vaccc19de_rki_data/main/data/decumulated_time_series.csv
}
