% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffs_data.R
\name{rki_download_diffs_ts}
\alias{rki_download_diffs_ts}
\title{download diffs time series from GitHub}
\usage{
rki_download_diffs_ts()
}
\value{
tibble. Diffs time series with the diffs (created by `rki_diffs`)
}
\description{
Downloads the diffs / 'decumulated' time series from the `vaccc19de_rki_data` repository, more specifically https://raw.githubusercontent.com/friep/vaccc19de_rki_data/main/data/decumulated_time_series.csv
}
