% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_raw_data.R
\name{rki_download_xlsx}
\alias{rki_download_xlsx}
\title{download excel file from rki website}
\usage{
rki_download_xlsx(out_folder = here::here())
}
\arguments{
\item{out_folder}{character. folder to store the excel file. defaults to here::here()}
}
\value{
character. path to excel file, invisibly.
}
\description{
uses rvest and curl to download the excel from rki
}
