% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{rki_extract_annotations}
\alias{rki_extract_annotations}
\title{extract annotations from raw data}
\usage{
rki_extract_annotations(.data)
}
\arguments{
\item{.data}{raw data frame of the sheet with the bundesland data}
}
\value{
data frame with Bundesland info and annotations
}
\description{
extract annotations
}
