% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffs_data.R
\name{rki_extract_diffs}
\alias{rki_extract_diffs}
\title{'decumulate' cumulative data to get diffs}
\usage{
rki_extract_diffs(cumulative_data)
}
\arguments{
\item{cumulative_data}{tibble. the cumulative time series.}
}
\value{
tibble. "decumulated" time series with the diffs, i.e. where values represent the difference since the previous update.
}
\description{
'decumulate' cumulative data to get diffs
}
