% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_raw_data.R
\name{rki_extract_sheet_csvs}
\alias{rki_extract_sheet_csvs}
\title{store sheets from xlsx as csvs}
\usage{
rki_extract_sheet_csvs(xlsx_path, out_folder = here::here())
}
\arguments{
\item{xlsx_path}{character. path to xlsx file}

\item{out_folder}{folder to store the csv files. defaults to here::here()}
}
\description{
store sheets from xlsx as csvs (for easier debugging).
}
