% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datenstand_timestamp.R
\name{rki_extract_ts_string}
\alias{rki_extract_ts_string}
\title{find the row with "Datenstand"
hopefully they don't change the format of this!!}
\usage{
rki_extract_ts_string(raw)
}
\description{
find the row with "Datenstand"
hopefully they don't change the format of this!!
}
