% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datenstand_timestamp.R
\name{rki_get_timestamp}
\alias{rki_get_timestamp}
\title{get timestamp of "Datenstand"}
\usage{
rki_get_timestamp(.data)
}
\arguments{
\item{.data}{raw data frame of sheet "Erläuterung"}
}
\value{
timestamp of Datenstand
}
\description{
get timestamp of "Datenstand"
}
