% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datenstand_timestamp.R
\name{rki_parse_ts}
\alias{rki_parse_ts}
\title{parse German timestamp with "Uhr" at the end - thankfully lubridate is clever like that}
\usage{
rki_parse_ts(s)
}
\description{
parse German timestamp with "Uhr" at the end - thankfully lubridate is clever like that
}
