% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{rki_tidyfy_header_dat}
\alias{rki_tidyfy_header_dat}
\title{clean data}
\usage{
rki_tidyfy_header_dat(xl, sheet)
}
\arguments{
\item{xl}{raw data frame}

\item{sheet}{"gesamt" or indikation"}
}
\value{
cleaned data in wide format
}
\description{
clean data
}
