% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_data.R
\name{rki_extract_cumulative_data}
\alias{rki_extract_cumulative_data}
\title{extract cumulative / original data}
\usage{
rki_extract_cumulative_data(xlsx_path)
}
\arguments{
\item{xlsx_path}{character. path to excel}
}
\value{
tibble. the data as displayed in sheet 2 enriched with the Datenstand timestamp from sheet 1.
}
\description{
returns the data as displayed in sheet 2 enriched with the Datenstand timestamp from sheet 1.
When collected over multiple days, this will result in cumulative data because the individual columns do not display the difference but total numbers
(except for column "differenz_zum_vortag").
}
